[TECH-272] Rename using single f
branch: feature/TECH-272-upgrade-monorepo
